nixpkgs-fmt all of the things

+++ pkgs/development/libraries/libinput/default.nix
@@ -1,29 +1,43 @@
-{ stdenv, fetchurl, pkgconfig, meson, ninja
-, libevdev, mtdev, udev, libwacom
-, documentationSupport ? false, doxygen ? null, graphviz ? null # Documentation
-, eventGUISupport ? false, cairo ? null, glib ? null, gtk3 ? null # GUI event viewer support
-, testsSupport ? false, check ? null, valgrind ? null, python3 ? null
+{ stdenv
+, fetchurl
+, pkgconfig
+, meson
+, ninja
+, libevdev
+, mtdev
+, udev
+, libwacom
+, documentationSupport ? false
+, doxygen ? null
+, graphviz ? null # Documentation
+, eventGUISupport ? false
+, cairo ? null
+, glib ? null
+, gtk3 ? null # GUI event viewer support
+, testsSupport ? false
+, check ? null
+, valgrind ? null
+, python3 ? null
 }:
 
 assert documentationSupport -> doxygen != null && graphviz != null && python3 != null;
 assert eventGUISupport -> cairo != null && glib != null && gtk3 != null;
 assert testsSupport -> check != null && valgrind != null && python3 != null;
-
 let
   mkFlag = optSet: flag: "-D${flag}=${stdenv.lib.boolToString optSet}";
 
-  sphinx-build = if documentationSupport then
-    python3.pkgs.sphinx.overrideAttrs (super: {
-      propagatedBuildInputs = super.propagatedBuildInputs ++ (with python3.pkgs; [ recommonmark sphinx_rtd_theme ]);
+  sphinx-build =
+    if documentationSupport then
+      python3.pkgs.sphinx.overrideAttrs (super: {
+        propagatedBuildInputs = super.propagatedBuildInputs ++ (with python3.pkgs; [ recommonmark sphinx_rtd_theme ]);
 
-      postFixup = super.postFixup or "" + ''
-        # Do not propagate Python
-        rm $out/nix-support/propagated-build-inputs
-      '';
-    })
-  else null;
+        postFixup = super.postFixup or "" + ''
+          # Do not propagate Python
+          rm $out/nix-support/propagated-build-inputs
+        '';
+      })
+    else null;
 in
-
 with stdenv.lib;
 stdenv.mkDerivation rec {
   pname = "libinput";
@@ -66,9 +80,9 @@ stdenv.mkDerivation rec {
 
   meta = {
     description = "Handles input devices in Wayland compositors and provides a generic X.Org input driver";
-    homepage    = http://www.freedesktop.org/wiki/Software/libinput;
-    license     = licenses.mit;
-    platforms   = platforms.unix;
+    homepage = http://www.freedesktop.org/wiki/Software/libinput;
+    license = licenses.mit;
+    platforms = platforms.unix;
     maintainers = with maintainers; [ codyopel ];
   };
 }
